Test Symfony2 indentation of switch-case statements via ERT
This patch adds a test for ‘tests/issue-111.php’ in our test suite, as
well as adding the necessary magic comments in the PHP script so that
the test knows what to expect.  The PHP script, however, indents the
statements after each case using the incorrect number of spaces, which
should be four spaces, as defined by PSR-2.

GitHub-Issue: 111